backend/requirements.txt
asgiref==3.7.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
coreapi==2.3.3
coreschema==0.0.4
cryptography==42.0.5
defusedxml==0.8.0rc2
diff-match-patch==20230430
Django==3.2.16
django-colorfield==0.9.0
django-cors-headers==3.13.0
django-filter==23.5
django-import-export==3.3.7
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-filter==0.2.2
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
drf-extra-fields==3.7.0
et-xmlfile==1.1.0
filetype==1.2.0
gunicorn==21.2.0
idna==3.6
iniconfig==2.0.0
itypes==1.2.0
Jinja2==3.1.3
MarkupPy==1.14
MarkupSafe==2.1.5
numpy==1.26.4
oauthlib==3.2.2
odfpy==1.4.1
openpyxl==3.1.2
packaging==24.0
pandas==2.2.1
Pillow==10.1.0
pluggy==1.4.0
psycopg2-binary==2.9.3
pycparser==2.21
PyJWT==2.8.0
pytest==8.1.1
pytest-django==4.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.4.0
six==1.16.0
social-auth-app-django==5.4.0
social-auth-core==4.5.3
sqlparse==0.4.4
subscribe==0.6.1
svgwrite==1.4.3
tablib==3.5.0
Tree==0.2.4
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
xlrd==2.0.1
xlwt==1.3.0
